move SelectedProfilesService into _shared module
